+++ pineappl_capi/src/lib.rs
@@ -567,7 +567,7 @@ pub unsafe extern "C" fn pineappl_grid_fill(
 ) {
     let grid = unsafe { &mut *grid };
 
-    grid.fill(order, observable, lumi, &[x1, x2, q2], weight);
+    grid.fill(order, observable, lumi, &[q2, x1, x2], weight);
 }
 
 /// Fill `grid` for the given momentum fractions `x1` and `x2`, at the scale `q2` for the given
@@ -592,7 +592,7 @@ pub unsafe extern "C" fn pineappl_grid_fill_all(
     let weights = unsafe { slice::from_raw_parts(weights, grid.channels().len()) };
 
     for (channel, &weight) in weights.iter().enumerate() {
-        grid.fill(order, observable, channel, &[x1, x2, q2], weight);
+        grid.fill(order, observable, channel, &[q2, x1, x2], weight);
     }
 }
 
@@ -627,7 +627,7 @@ pub unsafe extern "C" fn pineappl_grid_fill_array(
     for (&x1, &x2, &q2, &order, &observable, &lumi, &weight) in
         izip!(x1, x2, q2, orders, observables, lumis, weights)
     {
-        grid.fill(order, observable, lumi, &[x1, x2, q2], weight);
+        grid.fill(order, observable, lumi, &[q2, x1, x2], weight);
     }
 }